# api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: DAEMON Platform API v1
  version: 1.0.0
  description: |
    Multi-service HTTP API (platform-api, ontology-service, case-service, rules-engine, ingestion-service).
    Success responses use `{ data }`. Errors use `{ error: { code, message, requestId?, timestamp } }`.
servers:
  - url: http://127.0.0.1:8080
    description: platform-api
  - url: http://127.0.0.1:8081
    description: ontology-service
  - url: http://127.0.0.1:8084
    description: case-service
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        example: tenant-demo
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
paths:
...
  /v1/me:
    get:
      summary: Current user profile
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/TenantId" }]
      servers: [{ url: http://127.0.0.1:8080 }]
  /v1/audit/events:
    get:
      summary: List audit events
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/TenantId" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Offset" }
        - name: resourceType
          in: query
          schema: { type: string }
        - name: resourceId
          in: query
          schema: { type: string }
      servers: [{ url: http://127.0.0.1:8080 }]
    post:
      summary: Append audit event
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/TenantId" }]
      servers: [{ url: http://127.0.0.1:8080 }]
...